Force Cache-Control for all mime types
This should minimize the display errors in browser after software updates.
The subsequent ReGa process does not set the cache control header for all MIME types and so the CSS and JS files are cached in the browser.
This setting forces the use of this HTTP header and overwrites possible values.
After cleaning the cache once, the browser no longer caches the data and always reloads it.

Disclaimer:
Stability vs. speed?
I am aware, that this will increase the loading time, because now all files will be fetched fresh.
I have not checked any negative effects on existing external tools, add-ons or scripts.