Support for docker/container images (#1056)
This PR integrates a first-hand implementation for generating docker/container versions of RaspberryMatic. Please note, that this is still partly WIP and not released yet. So if you are testing the docker implementation you are somehwat on your own until we release this to the public. However, please warmly welcome this new supported virtual platform for RaspberryMatic! 

Co-authored-by: Jens Maus <mail@jens-maus.de>