Fix the loading of openvpn configs
Commit 107f3fb8ed19de2e3a781dac95b0f5c831a53f4e introduced a regression
in the /etc/init.d/S60openvpn script by adding a "./" to the command
listing the config files to load.
However, the command returns that "./" in the results that are iterated
as $NAME variables; the $NAME variable is then used in the generation of
the pid file path as '/var/run/openvpn.$NAME.pid', which becomes invalid.